refactor: Advice 변경
- Security 적용 후 필요 없어진 advice 정리
