+++ src/main/java/com/sedin/qna/advice/DuplicatedErrorAdvice.java
@@ -1,8 +1,8 @@
 package com.sedin.qna.advice;
 
-import com.sedin.qna.exception.DuplicatedException;
-import com.sedin.qna.network.ApiResponseCode;
-import com.sedin.qna.network.ApiResponseDto;
+import com.sedin.qna.common.exception.DuplicatedException;
+import com.sedin.qna.common.response.ApiResponseCode;
+import com.sedin.qna.common.response.ApiResponseDto;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.ResponseStatus;
